requirements.txt
grpcio>=1.30.0
matplotlib>=3.2.2
numpy>=1.19.0
protobuf>=3.12.2
psutil>=5.7.0
requests>=2.24.0
